requirements.txt
Django==5.1.2
gunicorn==23.0.0
dj-database-url==3.0.1
psycopg2-binary==2.9.11
whitenoise==6.11.0
django-cors-headers==4.4.0
django-taggit==6.1.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
asgiref==3.8.1
requests==2.32.3
python-dotenv==1.0.1
cloudinary==1.41.0
dj3-cloudinary-storage==0.0.6
django-storages[boto3]==1.14.4
boto3==1.35.44
